Build the 404 page for a lesson

Build the 404 page for a lesson

Usage

build_404(pkg, quiet = FALSE)

Arguments

pkg

a list object generated from pkgdown::as_pkgdown()

quiet

passed to build_html(). When FALSE (default), a message will be printed to the screen about the build progress. When TRUE, no messages are generated.

Value

TRUE if the page was successfully generated

Details

During the lesson build process, a 404 page with absolute links back to the source pages must be generated otherwise, subsequent attempts to escape the 404 page will be futile.

This function is intended to be run on a lesson website that has already been built and is called for its side-effect of creating a 404 page.

See also