Build a single episode html file

This is a Carpentries-specific wrapper around pkgdown::render_page() with templates from varnish. This function is largely for internal use and will likely change.

build_episode_html(
  path_md,
  path_src = NULL,
  page_back = "index.md",
  page_forward = "index.md",
  pkg,
  quiet = FALSE
)

Arguments

path_md
the path to the episode markdown (not RMarkdown) file (usually via build_episode_md()).

path_src
the default is NULL indicating that the source file should be determined from the sandpaper-source entry in the yaml header. If this is not present, then this option allows you to specify that file.

page_back
the URL for the previous page

page_forward
the URL for the next page

pkg
a pkgdown object containing metadata for the site

quiet
if TRUE, messages are not produced. Defaults to TRUE.

Value

TRUE if the page was successful, FALSE otherwise.

Note

this function is for internal use, but exported for those who know what they are doing.

See also

Examples

if (.Platform$OS.type == "windows") {
  options("sandpaper.use_renv" = FALSE)
}
if (!interactive() && getOption("sandpaper.use_renv")) {
  old <- renv::config$cache.symlinks()
  options(renv.config.cache.symlinks = FALSE)
  on.exit(options(renv.config.cache.symlinks = old), add = TRUE)
}
tmp <- tempfile()
create_lesson(tmp, open = FALSE)

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14j…

#> ✔ First episode created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14jA/file1de23fef6d1d/episodes/01-introduction.Rmd

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14j…

#> ℹ Workflows up-to-date!

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14j…

#> ✔ Lesson successfully created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14jA/file1de23fef6d1d

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14j…

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14jA/file1de23fef6d1d

suppressMessages(set_episodes(tmp, get_episodes(tmp), write = TRUE))
if (rmarkdown::pandoc_available("2.11")) {
  # we can only build this if we have pandoc
  build_lesson(tmp)
}

# create a new file in files
fun_file <- file.path(tmp, "episodes", "files", "fun.Rmd")
txt <- c(
 "---\ntitle: Fun times\n---\n\n",
 "# new page\n", 
 "This is coming from `r R.version.string`\n",
 "::: testimonial\n\n#### testimony!\n\nwhat\n:::\n"
)
file.create(fun_file)

#> [1] TRUE

writeLines(txt, fun_file)
hash <- tools::md5sum(fun_file)
res <- build_episode_md(fun_file, hash)

#> 
#> 
#> processing file: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpGt14jA/file1de23fef6d1d/episodes/files/fun.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |......................................................................| 100%
#>    inline R code fragments
#> 
#> 
#> output file: /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGt14jA/file1de23fef6d1d/site/built/fun.md
#> 
#> 

if (rmarkdown::pandoc_available("2.11")) {
  # we can only build this if we have pandoc
  build_episode_html(res, path_src = fun_file, 
    pkg = pkgdown::as_pkgdown(file.path(tmp, "site"))
  )
}