(INTERNAL) Build and deploy the site with continous integration

A platform-agnostic method of deploying a lesson website and cached content. This function takes advantage of git worktrees to render and push content to orphan branches, which can be used to host the lesson website. This function assumes that you are in a git clone from a remote and have write access to that remote.

Usage

ci_deploy(
  path = ".",
  md_branch = "md-outputs",
  site_branch = "gh-pages",
  remote = "origin",
  reset = FALSE
)

Arguments

path

path to the lesson

md_branch

the branch name that contains the markdown outputs

site_branch

the branch name that contains the full HTML site

remote

the name of the git remote to which you should deploy.

reset

if TRUE, the markdown cache is cleared before rebuilding, this defaults to FALSE meaning the markdown cache will be provisioned and used.

Value

Nothing, invisibly. This is used for it's side-effect

Details

ci_deploy() does the same thing as build_lesson(), except instead of storing the outputs under the site/ folder, it pushes the outputs to remote orphan branches (determined by the md_branch and site_branch arguments). These branches are used as the cache and the website, respectively. If these branches do not exist, they will be created.

Requirements

This function can only run in a non-interactive fashion. If you try to run it interactively, you will get an error message. It assumes that the following are true:

it is running in a script or automated workflow

it is running in a clone of a git repository

the remote exists and is writable

Unexpected consequences can arise from violating these assumptions.

Workflow

This function has a similar two-step workflow as build_lesson(), with a few extra steps to ensure that the git branches are set up correctly. Below are the steps with elements common to build_lesson() annotated with *

check that a git user and email is registered

* Validate the lesson and generate global variables with validate_lesson()

provision, build, and deploy markdown branch with ci_build_markdown() i. provision markdown branch with git_worktree_setup() ii. * build the markdown source documents with build_markdown() iii. commit and push the git worktree to the remote branch with github_worktree_commit()

provision, build, and deploy site branch with ci_build_site() i. provision site branch with git_worktree_setup() ii. * build the site HTML documents with build_site() iii. commit and push the git worktree to the remote branch with github_worktree_commit() iv. remove the git worktree with github_worktree_remove()

remove markdown git worktree with github_worktree_remove()

Note

this function is not for interactive use. It requires git to be installed on your machine and will destroy anything you have in the site/ folder. For R-based lessons it will rebuild all components if the lockfile has changed.

Examples

# For this example, we are setting up a temporary repository with a local
# remote called `sandpaper-local`. This demonstrates how `ci_deploy()`
# modifies the remote, but there are setup and teardown steps to run.
# The actual example is highlighted below under the DEPLOY comment.

# SETUP -------------------------------------------------------------------
snd <- asNamespace("sandpaper")
tik <- Sys.time()
cli::cli_h1("Set up")
#> 
#> ── Set up ──────────────────────────────────────────────────────────────
cli::cli_h2("Create Lesson")
#> 
#> ── Create Lesson ──
#> 
restore_fixture <- snd$create_test_lesson()
#> → Bootstrapping example lesson
#> → Searching for and installing available dependencies
#> * Discovering package dependencies ... Done!
#> The following packages were discovered:
#> 
#> # ~/work/_temp/Library ===============
#> - R6          2.5.1
#> - base64enc   0.1-3
#> - bslib       0.4.2
#> - cachem      1.0.7
#> - cli         3.6.0
#> - digest      0.6.31
#> - ellipsis    0.3.2
#> - evaluate    0.20
#> - fastmap     1.1.1
#> - fs          1.6.1
#> - glue        1.6.2
#> - highr       0.10
#> - htmltools   0.5.4
#> - jquerylib   0.1.4
#> - jsonlite    1.8.4
#> - knitr       1.42
#> - lifecycle   1.0.3
#> - magrittr    2.0.3
#> - memoise     2.0.1
#> - mime        0.12
#> - rappdirs    0.3.3
#> - rlang       1.0.6
#> - rmarkdown   2.20
#> - sass        0.4.5
#> - stringi     1.7.12
#> - stringr     1.5.0
#> - tinytex     0.44
#> - vctrs       0.5.2
#> - xfun        0.37
#> - yaml        2.3.7
#> 
#> They will be copied into the project library.
#> 
#> * Copying packages into the project library ... Done!
#> * Hydrated 30 packages in 0.23 seconds.
#> → Recording changes in lockfile
#> The version of R recorded in the lockfile will be updated:
#> - R           [* -> 4.2.2]
#> 
#> The following package(s) will be updated in the lockfile:
#> 
#> # GitHub =============================
#> - renv        [* -> rstudio/renv@HEAD]
#> 
#> # RSPM ===============================
#> - R6          [* -> 2.5.1]
#> - base64enc   [* -> 0.1-3]
#> - bslib       [* -> 0.4.2]
#> - cachem      [* -> 1.0.7]
#> - cli         [* -> 3.6.0]
#> - digest      [* -> 0.6.31]
#> - ellipsis    [* -> 0.3.2]
#> - evaluate    [* -> 0.20]
#> - fastmap     [* -> 1.1.1]
#> - fs          [* -> 1.6.1]
#> - glue        [* -> 1.6.2]
#> - highr       [* -> 0.10]
#> - htmltools   [* -> 0.5.4]
#> - jquerylib   [* -> 0.1.4]
#> - jsonlite    [* -> 1.8.4]
#> - knitr       [* -> 1.42]
#> - lifecycle   [* -> 1.0.3]
#> - magrittr    [* -> 2.0.3]
#> - memoise     [* -> 2.0.1]
#> - mime        [* -> 0.12]
#> - rappdirs    [* -> 0.3.3]
#> - rlang       [* -> 1.0.6]
#> - rmarkdown   [* -> 2.20]
#> - sass        [* -> 0.4.5]
#> - stringi     [* -> 1.7.12]
#> - stringr     [* -> 1.5.0]
#> - tinytex     [* -> 0.44]
#> - vctrs       [* -> 0.5.2]
#> - xfun        [* -> 0.37]
#> - yaml        [* -> 2.3.7]
#> 
#> * Lockfile written to '/tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/renv/profiles/lesson-requirements/renv.lock'.
res <- getOption("sandpaper.test_fixture")
cli::cli_h2("Create Remote")
#> 
#> ── Create Remote ──
#> 
rmt <- fs::file_temp(pattern = "REMOTE-")
snd$setup_local_remote(repo = res, remote = rmt, verbose = FALSE)
tok <- Sys.time()
cli::cli_alert_info("Elapsed time: {round(tok - tik, 2)} seconds")
#> ℹ Elapsed time: 3.21 seconds

# reporting -----
# The repository should only have one branch and the remote should be in
# sync with the local.
cli::cli_h2("Local status")
#> 
#> ── Local status ──
#> 
gert::git_branch_list(repo = res)[c('name', 'commit', 'updated')]
#> # A tibble: 2 × 3
#>   name                 commit                        updated            
#>   <chr>                <chr>                         <dttm>             
#> 1 main                 608f329366172344fbb48af044df… 2023-03-09 18:19:54
#> 2 sandpaper-local/main 608f329366172344fbb48af044df… 2023-03-09 18:19:54
cli::cli_h2("First episode status")
#> ── First episode status ──
#> 
gert::git_stat_files("episodes/introduction.Rmd", repo = res)
#> # A tibble: 1 × 5
#>   file             created             modified            commits head 
#> * <chr>            <dttm>              <dttm>                <int> <chr>
#> 1 episodes/introd… 2023-03-09 18:19:54 2023-03-09 18:19:54       1 608f…
gert::git_stat_files("episodes/introduction.Rmd", repo = rmt)
#> # A tibble: 1 × 5
#>   file             created             modified            commits head 
#> * <chr>            <dttm>              <dttm>                <int> <chr>
#> 1 episodes/introd… 2023-03-09 18:19:54 2023-03-09 18:19:54       1 608f…

# DEPLOY ------------------------------------------------------------------
tik <- Sys.time()
cli::cli_h1("deploy to remote")
#> ── deploy to remote ────────────────────────────────────────────────────
sandpaper:::ci_deploy(path = res, remote = "sandpaper-local")
#> ── Validating Fenced Divs ──────────────────────────────────────────────
#> ── Validating Internal Links and Images ────────────────────────────────
#> ::group::Create New Branch
#> Running git checkout --orphan md-outputs
#> Switched to a new branch 'md-outputs'
#> Running git rm -rf --quiet .
#> Running git commit --allow-empty -m 'Initializing md-outputs branch'
#> [md-outputs (root-commit) 6959ba9] Initializing md-outputs branch
#> Running git push sandpaper-local 'HEAD:md-outputs'
#> To /tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef
#>  * [new branch]      HEAD -> md-outputs
#> Running git checkout main
#> Your branch is up to date with 'sandpaper-local/main'.
#> Switched to branch 'main'
#> ::endgroup::
#> ::group::Fetch sandpaper-local/md-outputs
#> Running git remote set-branches sandpaper-local md-outputs
#> Running git fetch sandpaper-local md-outputs
#> From /tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef
#>  * branch            md-outputs -> FETCH_HEAD
#> Running git remote set-branches sandpaper-local '*'
#> ::endgroup::
#> ::group::Add worktree for sandpaper-local/md-outputs in site/built
#> Running git worktree add --track -B md-outputs \
#>   /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/site/built \
#>   sandpaper-local/md-outputs
#> Preparing worktree (resetting branch 'md-outputs'; was at 6959ba9)
#> branch 'md-outputs' set up to track 'sandpaper-local/md-outputs'.
#> HEAD is now at 6959ba9 Initializing md-outputs branch
#> ::endgroup::
#> ::group::Build Markdown Sources
#> ℹ Consent to use package cache provided
#> → Searching for and installing available dependencies
#> Finding R package dependencies ... Done!
#> → Restoring any dependency versions
#> * The library is already synchronized with the lockfile.
#> → Recording changes in lockfile
#> * The lockfile is already up to date.
#> ℹ Using package cache in /home/runner/.cache/R/renv
#> Diagnostics Report [renv 0.17.0-39]
#> ===================================
#> 
#> # Session Info =======================
#> R version 4.2.2 (2022-10-31)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Ubuntu 22.04.2 LTS
#> 
#> Matrix products: default
#> BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3
#> LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so
#> 
#> locale:
#>  [1] LC_CTYPE=C.UTF-8       LC_NUMERIC=C           LC_TIME=C.UTF-8       
#>  [4] LC_COLLATE=C           LC_MONETARY=C.UTF-8    LC_MESSAGES=C.UTF-8   
#>  [7] LC_PAPER=C.UTF-8       LC_NAME=C              LC_ADDRESS=C          
#> [10] LC_TELEPHONE=C         LC_MEASUREMENT=C.UTF-8 LC_IDENTIFICATION=C   
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> loaded via a namespace (and not attached):
#> [1] compiler_4.2.2 tools_4.2.2    renv_0.17.0-39
#> 
#> # Project ============================
#> Project path: "/tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example"
#> 
#> # Status =============================
#> * The project is already synchronized with the lockfile.
#> 
#> # Packages ===========================
#>              Library Source  Lockfile Source Path Dependency
#> KernSmooth   2.23-20   CRAN      <NA>   <NA>  [2]       <NA>
#> MASS        7.3-58.1   CRAN      <NA>   <NA>  [2]       <NA>
#> Matrix         1.5-1   CRAN      <NA>   <NA>  [2]       <NA>
#> R6             2.5.1   RSPM     2.5.1   RSPM  [1]   indirect
#> base64enc      0.1-3   RSPM     0.1-3   RSPM  [1]   indirect
#> boot          1.3-28   CRAN      <NA>   <NA>  [2]       <NA>
#> bslib          0.4.2   RSPM     0.4.2   RSPM  [1]   indirect
#> cachem         1.0.7   RSPM     1.0.7   RSPM  [1]   indirect
#> class         7.3-20   CRAN      <NA>   <NA>  [2]       <NA>
#> cli            3.6.0   RSPM     3.6.0   RSPM  [1]   indirect
#> cluster        2.1.4   CRAN      <NA>   <NA>  [2]       <NA>
#> codetools     0.2-18   CRAN      <NA>   <NA>  [2]       <NA>
#> digest        0.6.31   RSPM    0.6.31   RSPM  [1]   indirect
#> ellipsis       0.3.2   RSPM     0.3.2   RSPM  [1]   indirect
#> evaluate        0.20   RSPM      0.20   RSPM  [1]   indirect
#> fastmap        1.1.1   RSPM     1.1.1   RSPM  [1]   indirect
#> foreign       0.8-83   CRAN      <NA>   <NA>  [2]       <NA>
#> fs             1.6.1   RSPM     1.6.1   RSPM  [1]   indirect
#> glue           1.6.2   RSPM     1.6.2   RSPM  [1]   indirect
#> grDevices       <NA>   <NA>      <NA>   <NA>  [2]   indirect
#> graphics        <NA>   <NA>      <NA>   <NA>  [2]   indirect
#> highr           0.10   RSPM      0.10   RSPM  [1]   indirect
#> htmltools      0.5.4   RSPM     0.5.4   RSPM  [1]   indirect
#> jquerylib      0.1.4   RSPM     0.1.4   RSPM  [1]   indirect
#> jsonlite       1.8.4   RSPM     1.8.4   RSPM  [1]   indirect
#> knitr           1.42   RSPM      1.42   RSPM  [1]   indirect
#> lattice      0.20-45   CRAN      <NA>   <NA>  [2]       <NA>
#> lifecycle      1.0.3   RSPM     1.0.3   RSPM  [1]   indirect
#> magrittr       2.0.3   RSPM     2.0.3   RSPM  [1]   indirect
#> memoise        2.0.1   RSPM     2.0.1   RSPM  [1]   indirect
#> methods         <NA>   <NA>      <NA>   <NA>  [2]   indirect
#> mgcv          1.8-41   CRAN      <NA>   <NA>  [2]       <NA>
#> mime            0.12   RSPM      0.12   RSPM  [1]   indirect
#> nlme         3.1-160   CRAN      <NA>   <NA>  [2]       <NA>
#> nnet          7.3-18   CRAN      <NA>   <NA>  [2]       <NA>
#> rappdirs       0.3.3   RSPM     0.3.3   RSPM  [1]   indirect
#> renv       0.17.0-39 GitHub 0.17.0-39 GitHub  [1]       <NA>
#> rlang          1.0.6   RSPM     1.0.6   RSPM  [1]   indirect
#> rmarkdown       2.20   RSPM      2.20   RSPM  [1]     direct
#> rpart         4.1.19   CRAN      <NA>   <NA>  [2]       <NA>
#> sass           0.4.5   RSPM     0.4.5   RSPM  [1]   indirect
#> spatial       7.3-15   CRAN      <NA>   <NA>  [2]       <NA>
#> stats           <NA>   <NA>      <NA>   <NA>  [2]   indirect
#> stringi       1.7.12   RSPM    1.7.12   RSPM  [1]   indirect
#> stringr        1.5.0   RSPM     1.5.0   RSPM  [1]   indirect
#> survival       3.4-0   CRAN      <NA>   <NA>  [2]       <NA>
#> tinytex         0.44   RSPM      0.44   RSPM  [1]   indirect
#> tools           <NA>   <NA>      <NA>   <NA>  [2]   indirect
#> utils           <NA>   <NA>      <NA>   <NA>  [2]   indirect
#> vctrs          0.5.2   RSPM     0.5.2   RSPM  [1]   indirect
#> xfun            0.37   RSPM      0.37   RSPM  [1]   indirect
#> yaml           2.3.7   RSPM     2.3.7   RSPM  [1]   indirect
#> 
#> [1]: /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/renv/profiles/lesson-requirements/renv/library/R-4.2/x86_64-pc-linux-gnu
#> [2]: /home/runner/.cache/R/renv/sandbox/R-4.2/x86_64-pc-linux-gnu/99f2a2ef                                                   
#> 
#> # ABI ================================
#> * No ABI conflicts were detected in the set of installed packages.
#> 
#> # User Profile =======================
#> [1] Source  Package Require Version Dev    
#> <0 rows> (or 0-length row.names)
#> 
#> # Settings ===========================
#> List of 10
#>  $ bioconductor.version     : chr(0) 
#>  $ external.libraries       : chr(0) 
#>  $ ignored.packages         : chr(0) 
#>  $ package.dependency.fields: chr [1:3] "Imports" "Depends" "LinkingTo"
#>  $ r.version                : chr(0) 
#>  $ snapshot.type            : chr "implicit"
#>  $ use.cache                : logi TRUE
#>  $ vcs.ignore.cellar        : logi TRUE
#>  $ vcs.ignore.library       : logi TRUE
#>  $ vcs.ignore.local         : logi TRUE
#> 
#> # Options ============================
#> List of 8
#>  $ defaultPackages                     : chr [1:6] "datasets" "utils" "grDevices" "graphics" ...
#>  $ download.file.method                : NULL
#>  $ download.file.extra                 : NULL
#>  $ install.packages.compile.from.source: NULL
#>  $ pkgType                             : chr "source"
#>  $ repos                               : Named chr [1:3] "https://carpentries.r-universe.dev" "https://carpentries.github.io/drat" "https://cran.rstudio.com"
#>   ..- attr(*, "names")= chr [1:3] "carpentries" "carpentries_archive" "CRAN"
#>  $ renv.project.path                   : chr "/tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example"
#>  $ renv.verbose                        : logi TRUE
#> 
#> # Environment Variables ==============
#> HOME                        = /home/runner
#> LANG                        = C.UTF-8
#> MAKE                        = make
#> RENV_CONFIG_CACHE_SYMLINKS  = FALSE
#> RENV_CONFIG_REPOS_OVERRIDE  = https://packagemanager.posit.co/cran/__linux__/jammy/latest
#> RENV_DEFAULT_R_ENVIRON      = <NA>
#> RENV_DEFAULT_R_ENVIRON_USER = <NA>
#> RENV_DEFAULT_R_LIBS         = <NA>
#> RENV_DEFAULT_R_LIBS_SITE    = /opt/R/4.2.2/lib/R/site-library
#> RENV_DEFAULT_R_LIBS_USER    = /home/runner/work/_temp/Library
#> RENV_DEFAULT_R_PROFILE      = <NA>
#> RENV_DEFAULT_R_PROFILE_USER = <NA>
#> RENV_PROFILE                = lesson-requirements
#> RENV_PROJECT                = /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example
#> R_LIBS                      = <NA>
#> R_LIBS_SITE                 = /opt/R/4.2.2/lib/R/site-library
#> R_LIBS_USER                 = /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/renv/profiles/lesson-requirements/renv/library/R-4.2/x86_64-pc-linux-gnu:/home/runner/.cache/R/renv/sandbox/R-4.2/x86_64-pc-linux-gnu/99f2a2ef
#> 
#> # PATH ===============================
#> - /home/runner/.local/bin
#> - /opt/pipx_bin
#> - /home/runner/.cargo/bin
#> - /home/runner/.config/composer/vendor/bin
#> - /usr/local/.ghcup/bin
#> - /home/runner/.dotnet/tools
#> - /snap/bin
#> - /usr/local/sbin
#> - /usr/local/bin
#> - /usr/sbin
#> - /usr/bin
#> - /sbin
#> - /bin
#> - /usr/games
#> - /usr/local/games
#> - /snap/bin
#> 
#> # Cache ==============================
#> There are a total of 0 package(s) installed in the renv cache.
#> Cache path: "~/.cache/R/renv/cache/v5/R-4.2/x86_64-pc-linux-gnu"
#> 
#> 
#> 
#> processing file: /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/episodes/introduction.Rmd
#>   |                                                                    |                                                            |   0%  |                                                                    |....................                                        |  33%            |                                                                    |........................................                    |  67% (pyramid)  |                                                                    |............................................................| 100%                                                                                                              
#> output file: /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/site/built/introduction.md
#> 
#> ::endgroup::
#> ::group::Commit Markdown Sources
#> Running git commit --allow-empty -m \
#>   'markdown source builds
#> 
#> Auto-generated via {sandpaper}
#> Source  : 608f329366172344fbb48af044df4a32b4673549
#> Branch  : main
#> Author  : Jerry Johnson <jerry@gmail.com>
#> Time    : 2023-03-09 18:19:54 +0000
#> Message : Initial commit [via {sandpaper}]
#> '
#> [md-outputs 1891d8c] markdown source builds
#>  12 files changed, 958 insertions(+)
#>  create mode 100644 CODE_OF_CONDUCT.md
#>  create mode 100644 LICENSE.md
#>  create mode 100644 config.yaml
#>  create mode 100644 fig/introduction-rendered-pyramid-1.png
#>  create mode 100644 index.md
#>  create mode 100644 instructor-notes.md
#>  create mode 100644 introduction.md
#>  create mode 100644 learner-profiles.md
#>  create mode 100644 links.md
#>  create mode 100644 md5sum.txt
#>  create mode 100644 renv.lock
#>  create mode 100644 setup.md
#> Running git remote -v
#> sandpaper-local	/tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef (fetch)
#> sandpaper-local	/tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef (push)
#> Running git push --force sandpaper-local 'HEAD:md-outputs'
#> To /tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef
#>    6959ba9..1891d8c  HEAD -> md-outputs
#> ::endgroup::
#> ::group::Create New Branch
#> Running git checkout --orphan gh-pages
#> Switched to a new branch 'gh-pages'
#> Running git rm -rf --quiet .
#> Running git commit --allow-empty -m 'Initializing gh-pages branch'
#> [gh-pages (root-commit) 14fa33c] Initializing gh-pages branch
#> Running git push sandpaper-local 'HEAD:gh-pages'
#> To /tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef
#>  * [new branch]      HEAD -> gh-pages
#> Running git checkout main
#> Your branch is up to date with 'sandpaper-local/main'.
#> Switched to branch 'main'
#> ::endgroup::
#> ::group::Fetch sandpaper-local/gh-pages
#> Running git remote set-branches sandpaper-local gh-pages
#> Running git fetch sandpaper-local gh-pages
#> From /tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef
#>  * branch            gh-pages   -> FETCH_HEAD
#> Running git remote set-branches sandpaper-local '*'
#> ::endgroup::
#> ::group::Add worktree for sandpaper-local/gh-pages in site/docs
#> Running git worktree add --track -B gh-pages \
#>   /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/site/docs \
#>   sandpaper-local/gh-pages
#> Preparing worktree (resetting branch 'gh-pages'; was at 14fa33c)
#> branch 'gh-pages' set up to track 'sandpaper-local/gh-pages'.
#> HEAD is now at 14fa33c Initializing gh-pages branch
#> ::endgroup::
#> ::group::Build Lesson Website
#> ◉ pandoc found
#>   version : 2.19.2
#>   path    : /usr/bin
#> -- Initialising site ---------------------------------------------------
#> Copying '../../../../../home/runner/work/_temp/Library/pkgdown/BS3/assets/bootstrap-toc.css' to 'bootstrap-toc.css'
#> Copying '../../../../../home/runner/work/_temp/Library/pkgdown/BS3/assets/bootstrap-toc.js' to 'bootstrap-toc.js'
#> Copying '../../../../../home/runner/work/_temp/Library/pkgdown/BS3/assets/docsearch.css' to 'docsearch.css'
#> Copying '../../../../../home/runner/work/_temp/Library/pkgdown/BS3/assets/docsearch.js' to 'docsearch.js'
#> Copying '../../../../../home/runner/work/_temp/Library/pkgdown/BS3/assets/link.svg' to 'link.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/pkgdown/BS3/assets/pkgdown.css' to 'pkgdown.css'
#> Copying '../../../../../home/runner/work/_temp/Library/pkgdown/BS3/assets/pkgdown.js' to 'pkgdown.js'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/android-chrome-192x192.png' to 'android-chrome-192x192.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/android-chrome-512x512.png' to 'android-chrome-512x512.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/apple-touch-icon.png' to 'apple-touch-icon.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/Mulish-Bold.ttf' to 'assets/fonts/Mulish-Bold.ttf'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/Mulish-Bold.woff' to 'assets/fonts/Mulish-Bold.woff'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/Mulish-ExtraBold.ttf' to 'assets/fonts/Mulish-ExtraBold.ttf'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/mulish-v5-latin-regular.eot' to 'assets/fonts/mulish-v5-latin-regular.eot'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/mulish-v5-latin-regular.svg' to 'assets/fonts/mulish-v5-latin-regular.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/mulish-v5-latin-regular.ttf' to 'assets/fonts/mulish-v5-latin-regular.ttf'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/mulish-v5-latin-regular.woff' to 'assets/fonts/mulish-v5-latin-regular.woff'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/mulish-v5-latin-regular.woff2' to 'assets/fonts/mulish-v5-latin-regular.woff2'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/mulish-variablefont_wght.woff' to 'assets/fonts/mulish-variablefont_wght.woff'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/fonts/mulish-variablefont_wght.woff2' to 'assets/fonts/mulish-variablefont_wght.woff2'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/carpentries-logo-sm.svg' to 'assets/images/carpentries-logo-sm.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/carpentries-logo.svg' to 'assets/images/carpentries-logo.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/data-logo-sm.svg' to 'assets/images/data-logo-sm.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/data-logo.svg' to 'assets/images/data-logo.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/dropdown-arrow.svg' to 'assets/images/dropdown-arrow.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/incubator-logo-sm.svg' to 'assets/images/incubator-logo-sm.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/incubator-logo.svg' to 'assets/images/incubator-logo.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/lab-logo-sm.svg' to 'assets/images/lab-logo-sm.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/lab-logo.svg' to 'assets/images/lab-logo.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/library-logo-sm.svg' to 'assets/images/library-logo-sm.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/library-logo.svg' to 'assets/images/library-logo.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/minus.svg' to 'assets/images/minus.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/plus.svg' to 'assets/images/plus.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/software-logo-sm.svg' to 'assets/images/software-logo-sm.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/images/software-logo.svg' to 'assets/images/software-logo.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/scripts.js' to 'assets/scripts.js'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/styles.css' to 'assets/styles.css'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/assets/styles.css.map' to 'assets/styles.css.map'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicon-16x16.png' to 'favicon-16x16.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicon-32x32.png' to 'favicon-32x32.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-114x114.png' to 'favicons/cp/apple-touch-icon-114x114.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-120x120.png' to 'favicons/cp/apple-touch-icon-120x120.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-144x144.png' to 'favicons/cp/apple-touch-icon-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-152x152.png' to 'favicons/cp/apple-touch-icon-152x152.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-57x57.png' to 'favicons/cp/apple-touch-icon-57x57.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-60x60.png' to 'favicons/cp/apple-touch-icon-60x60.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-72x72.png' to 'favicons/cp/apple-touch-icon-72x72.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/apple-touch-icon-76x76.png' to 'favicons/cp/apple-touch-icon-76x76.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/favicon-128.png' to 'favicons/cp/favicon-128.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/favicon-16x16.png' to 'favicons/cp/favicon-16x16.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/favicon-196x196.png' to 'favicons/cp/favicon-196x196.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/favicon-32x32.png' to 'favicons/cp/favicon-32x32.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/favicon-96x96.png' to 'favicons/cp/favicon-96x96.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/favicon.ico' to 'favicons/cp/favicon.ico'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/mstile-144x144.png' to 'favicons/cp/mstile-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/mstile-150x150.png' to 'favicons/cp/mstile-150x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/mstile-310x150.png' to 'favicons/cp/mstile-310x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/mstile-310x310.png' to 'favicons/cp/mstile-310x310.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/cp/mstile-70x70.png' to 'favicons/cp/mstile-70x70.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-114x114.png' to 'favicons/dc/apple-touch-icon-114x114.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-120x120.png' to 'favicons/dc/apple-touch-icon-120x120.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-144x144.png' to 'favicons/dc/apple-touch-icon-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-152x152.png' to 'favicons/dc/apple-touch-icon-152x152.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-57x57.png' to 'favicons/dc/apple-touch-icon-57x57.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-60x60.png' to 'favicons/dc/apple-touch-icon-60x60.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-72x72.png' to 'favicons/dc/apple-touch-icon-72x72.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/apple-touch-icon-76x76.png' to 'favicons/dc/apple-touch-icon-76x76.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/favicon-128.png' to 'favicons/dc/favicon-128.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/favicon-16x16.png' to 'favicons/dc/favicon-16x16.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/favicon-196x196.png' to 'favicons/dc/favicon-196x196.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/favicon-32x32.png' to 'favicons/dc/favicon-32x32.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/favicon-96x96.png' to 'favicons/dc/favicon-96x96.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/favicon.ico' to 'favicons/dc/favicon.ico'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/mstile-144x144.png' to 'favicons/dc/mstile-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/mstile-150x150.png' to 'favicons/dc/mstile-150x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/mstile-310x150.png' to 'favicons/dc/mstile-310x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/mstile-310x310.png' to 'favicons/dc/mstile-310x310.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/dc/mstile-70x70.png' to 'favicons/dc/mstile-70x70.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-114x114.png' to 'favicons/lc/apple-touch-icon-114x114.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-120x120.png' to 'favicons/lc/apple-touch-icon-120x120.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-144x144.png' to 'favicons/lc/apple-touch-icon-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-152x152.png' to 'favicons/lc/apple-touch-icon-152x152.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-57x57.png' to 'favicons/lc/apple-touch-icon-57x57.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-60x60.png' to 'favicons/lc/apple-touch-icon-60x60.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-72x72.png' to 'favicons/lc/apple-touch-icon-72x72.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/apple-touch-icon-76x76.png' to 'favicons/lc/apple-touch-icon-76x76.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/favicon-128.png' to 'favicons/lc/favicon-128.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/favicon-16x16.png' to 'favicons/lc/favicon-16x16.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/favicon-196x196.png' to 'favicons/lc/favicon-196x196.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/favicon-32x32.png' to 'favicons/lc/favicon-32x32.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/favicon-96x96.png' to 'favicons/lc/favicon-96x96.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/favicon.ico' to 'favicons/lc/favicon.ico'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/mstile-144x144.png' to 'favicons/lc/mstile-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/mstile-150x150.png' to 'favicons/lc/mstile-150x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/mstile-310x150.png' to 'favicons/lc/mstile-310x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/mstile-310x310.png' to 'favicons/lc/mstile-310x310.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/lc/mstile-70x70.png' to 'favicons/lc/mstile-70x70.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-114x114.png' to 'favicons/swc/apple-touch-icon-114x114.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-120x120.png' to 'favicons/swc/apple-touch-icon-120x120.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-144x144.png' to 'favicons/swc/apple-touch-icon-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-152x152.png' to 'favicons/swc/apple-touch-icon-152x152.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-57x57.png' to 'favicons/swc/apple-touch-icon-57x57.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-60x60.png' to 'favicons/swc/apple-touch-icon-60x60.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-72x72.png' to 'favicons/swc/apple-touch-icon-72x72.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/apple-touch-icon-76x76.png' to 'favicons/swc/apple-touch-icon-76x76.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/favicon-128.png' to 'favicons/swc/favicon-128.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/favicon-16x16.png' to 'favicons/swc/favicon-16x16.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/favicon-196x196.png' to 'favicons/swc/favicon-196x196.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/favicon-32x32.png' to 'favicons/swc/favicon-32x32.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/favicon-96x96.png' to 'favicons/swc/favicon-96x96.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/favicon.ico' to 'favicons/swc/favicon.ico'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/mstile-144x144.png' to 'favicons/swc/mstile-144x144.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/mstile-150x150.png' to 'favicons/swc/mstile-150x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/mstile-310x150.png' to 'favicons/swc/mstile-310x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/mstile-310x310.png' to 'favicons/swc/mstile-310x310.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/favicons/swc/mstile-70x70.png' to 'favicons/swc/mstile-70x70.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/mstile-150x150.png' to 'mstile-150x150.png'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/safari-pinned-tab.svg' to 'safari-pinned-tab.svg'
#> Copying '../../../../../home/runner/work/_temp/Library/varnish/pkgdown/assets/site.webmanifest' to 'site.webmanifest'
#> ── Scanning episodes to rebuild ────────────────────────────────────────
#> Writing 'instructor/CODE_OF_CONDUCT.html'
#> Writing 'CODE_OF_CONDUCT.html'
#> Writing 'instructor/LICENSE.html'
#> Writing 'LICENSE.html'
#> Writing 'instructor/introduction.html'
#> Writing 'introduction.html'
#> ── Creating learner profiles ───────────────────────────────────────────
#> Writing 'instructor/profiles.html'
#> Writing 'profiles.html'
#> ── Creating homepage ───────────────────────────────────────────────────
#> Writing 'instructor/index.html'
#> Writing 'index.html'
#> ── Creating keypoints summary ──────────────────────────────────────────
#> Writing 'instructor/key-points.html'
#> Writing 'key-points.html'
#> ── Creating All-in-one page ────────────────────────────────────────────
#> Writing 'instructor/aio.html'
#> Writing 'aio.html'
#> ── Creating Images page ────────────────────────────────────────────────
#> Writing 'instructor/images.html'
#> Writing 'images.html'
#> ── Creating Instructor Notes ───────────────────────────────────────────
#> Writing 'instructor/instructor-notes.html'
#> Writing 'instructor-notes.html'
#> ── Creating sitemap.xml ────────────────────────────────────────────────
#> 
#> Output created: /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/site/docs/index.html
#> ::endgroup::
#> ::group::Commit Lesson Website
#> Running git commit --allow-empty -m \
#>   'site deploy
#> 
#> Auto-generated via {sandpaper}
#> Source  : 1891d8ca88f81c4117328db2607e9e4e2c1965fb
#> Branch  : md-outputs
#> Author  : Jerry Johnson <jerry@gmail.com>
#> Time    : 2023-03-09 18:20:00 +0000
#> Message : markdown source builds
#> 
#> Auto-generated via {sandpaper}
#> Source  : 608f329366172344fbb48af044df4a32b4673549
#> Branch  : main
#> Author  : Jerry Johnson <jerry@gmail.com>
#> Time    : 2023-03-09 18:19:54 +0000
#> Message : Initial commit [via {sandpaper}]
#> '
#> [gh-pages cd35bcf] site deploy
#>  144 files changed, 9025 insertions(+)
#>  create mode 100644 .nojekyll
#>  create mode 100644 CODE_OF_CONDUCT.html
#>  create mode 100644 LICENSE.html
#>  create mode 100644 aio.html
#>  create mode 100644 android-chrome-192x192.png
#>  create mode 100644 android-chrome-512x512.png
#>  create mode 100644 apple-touch-icon.png
#>  create mode 100644 assets/fonts/Mulish-Bold.ttf
#>  create mode 100644 assets/fonts/Mulish-Bold.woff
#>  create mode 100644 assets/fonts/Mulish-ExtraBold.ttf
#>  create mode 100644 assets/fonts/mulish-v5-latin-regular.eot
#>  create mode 100644 assets/fonts/mulish-v5-latin-regular.svg
#>  create mode 100644 assets/fonts/mulish-v5-latin-regular.ttf
#>  create mode 100644 assets/fonts/mulish-v5-latin-regular.woff
#>  create mode 100644 assets/fonts/mulish-v5-latin-regular.woff2
#>  create mode 100644 assets/fonts/mulish-variablefont_wght.woff
#>  create mode 100644 assets/fonts/mulish-variablefont_wght.woff2
#>  create mode 100644 assets/images/carpentries-logo-sm.svg
#>  create mode 100644 assets/images/carpentries-logo.svg
#>  create mode 100644 assets/images/data-logo-sm.svg
#>  create mode 100644 assets/images/data-logo.svg
#>  create mode 100644 assets/images/dropdown-arrow.svg
#>  create mode 100644 assets/images/incubator-logo-sm.svg
#>  create mode 100644 assets/images/incubator-logo.svg
#>  create mode 100644 assets/images/lab-logo-sm.svg
#>  create mode 100644 assets/images/lab-logo.svg
#>  create mode 100644 assets/images/library-logo-sm.svg
#>  create mode 100644 assets/images/library-logo.svg
#>  create mode 100644 assets/images/minus.svg
#>  create mode 100644 assets/images/plus.svg
#>  create mode 100644 assets/images/software-logo-sm.svg
#>  create mode 100644 assets/images/software-logo.svg
#>  create mode 100644 assets/scripts.js
#>  create mode 100644 assets/styles.css
#>  create mode 100644 assets/styles.css.map
#>  create mode 100644 bootstrap-toc.css
#>  create mode 100644 bootstrap-toc.js
#>  create mode 100644 config.yaml
#>  create mode 100644 docsearch.css
#>  create mode 100644 docsearch.js
#>  create mode 100644 favicon-16x16.png
#>  create mode 100644 favicon-32x32.png
#>  create mode 100644 favicons/cp/apple-touch-icon-114x114.png
#>  create mode 100644 favicons/cp/apple-touch-icon-120x120.png
#>  create mode 100644 favicons/cp/apple-touch-icon-144x144.png
#>  create mode 100644 favicons/cp/apple-touch-icon-152x152.png
#>  create mode 100644 favicons/cp/apple-touch-icon-57x57.png
#>  create mode 100644 favicons/cp/apple-touch-icon-60x60.png
#>  create mode 100644 favicons/cp/apple-touch-icon-72x72.png
#>  create mode 100644 favicons/cp/apple-touch-icon-76x76.png
#>  create mode 100644 favicons/cp/favicon-128.png
#>  create mode 100644 favicons/cp/favicon-16x16.png
#>  create mode 100644 favicons/cp/favicon-196x196.png
#>  create mode 100644 favicons/cp/favicon-32x32.png
#>  create mode 100644 favicons/cp/favicon-96x96.png
#>  create mode 100644 favicons/cp/favicon.ico
#>  create mode 100644 favicons/cp/mstile-144x144.png
#>  create mode 100644 favicons/cp/mstile-150x150.png
#>  create mode 100644 favicons/cp/mstile-310x150.png
#>  create mode 100644 favicons/cp/mstile-310x310.png
#>  create mode 100644 favicons/cp/mstile-70x70.png
#>  create mode 100644 favicons/dc/apple-touch-icon-114x114.png
#>  create mode 100644 favicons/dc/apple-touch-icon-120x120.png
#>  create mode 100644 favicons/dc/apple-touch-icon-144x144.png
#>  create mode 100644 favicons/dc/apple-touch-icon-152x152.png
#>  create mode 100644 favicons/dc/apple-touch-icon-57x57.png
#>  create mode 100644 favicons/dc/apple-touch-icon-60x60.png
#>  create mode 100644 favicons/dc/apple-touch-icon-72x72.png
#>  create mode 100644 favicons/dc/apple-touch-icon-76x76.png
#>  create mode 100644 favicons/dc/favicon-128.png
#>  create mode 100644 favicons/dc/favicon-16x16.png
#>  create mode 100644 favicons/dc/favicon-196x196.png
#>  create mode 100644 favicons/dc/favicon-32x32.png
#>  create mode 100644 favicons/dc/favicon-96x96.png
#>  create mode 100644 favicons/dc/favicon.ico
#>  create mode 100644 favicons/dc/mstile-144x144.png
#>  create mode 100644 favicons/dc/mstile-150x150.png
#>  create mode 100644 favicons/dc/mstile-310x150.png
#>  create mode 100644 favicons/dc/mstile-310x310.png
#>  create mode 100644 favicons/dc/mstile-70x70.png
#>  create mode 100644 favicons/lc/apple-touch-icon-114x114.png
#>  create mode 100644 favicons/lc/apple-touch-icon-120x120.png
#>  create mode 100644 favicons/lc/apple-touch-icon-144x144.png
#>  create mode 100644 favicons/lc/apple-touch-icon-152x152.png
#>  create mode 100644 favicons/lc/apple-touch-icon-57x57.png
#>  create mode 100644 favicons/lc/apple-touch-icon-60x60.png
#>  create mode 100644 favicons/lc/apple-touch-icon-72x72.png
#>  create mode 100644 favicons/lc/apple-touch-icon-76x76.png
#>  create mode 100644 favicons/lc/favicon-128.png
#>  create mode 100644 favicons/lc/favicon-16x16.png
#>  create mode 100644 favicons/lc/favicon-196x196.png
#>  create mode 100644 favicons/lc/favicon-32x32.png
#>  create mode 100644 favicons/lc/favicon-96x96.png
#>  create mode 100644 favicons/lc/favicon.ico
#>  create mode 100644 favicons/lc/mstile-144x144.png
#>  create mode 100644 favicons/lc/mstile-150x150.png
#>  create mode 100644 favicons/lc/mstile-310x150.png
#>  create mode 100644 favicons/lc/mstile-310x310.png
#>  create mode 100644 favicons/lc/mstile-70x70.png
#>  create mode 100644 favicons/swc/apple-touch-icon-114x114.png
#>  create mode 100644 favicons/swc/apple-touch-icon-120x120.png
#>  create mode 100644 favicons/swc/apple-touch-icon-144x144.png
#>  create mode 100644 favicons/swc/apple-touch-icon-152x152.png
#>  create mode 100644 favicons/swc/apple-touch-icon-57x57.png
#>  create mode 100644 favicons/swc/apple-touch-icon-60x60.png
#>  create mode 100644 favicons/swc/apple-touch-icon-72x72.png
#>  create mode 100644 favicons/swc/apple-touch-icon-76x76.png
#>  create mode 100644 favicons/swc/favicon-128.png
#>  create mode 100644 favicons/swc/favicon-16x16.png
#>  create mode 100644 favicons/swc/favicon-196x196.png
#>  create mode 100644 favicons/swc/favicon-32x32.png
#>  create mode 100644 favicons/swc/favicon-96x96.png
#>  create mode 100644 favicons/swc/favicon.ico
#>  create mode 100644 favicons/swc/mstile-144x144.png
#>  create mode 100644 favicons/swc/mstile-150x150.png
#>  create mode 100644 favicons/swc/mstile-310x150.png
#>  create mode 100644 favicons/swc/mstile-310x310.png
#>  create mode 100644 favicons/swc/mstile-70x70.png
#>  create mode 100644 fig/introduction-rendered-pyramid-1.png
#>  create mode 100644 images.html
#>  create mode 100644 index.html
#>  create mode 100644 instructor-notes.html
#>  create mode 100644 instructor/CODE_OF_CONDUCT.html
#>  create mode 100644 instructor/LICENSE.html
#>  create mode 100644 instructor/aio.html
#>  create mode 100644 instructor/images.html
#>  create mode 100644 instructor/index.html
#>  create mode 100644 instructor/instructor-notes.html
#>  create mode 100644 instructor/introduction.html
#>  create mode 100644 instructor/key-points.html
#>  create mode 100644 instructor/profiles.html
#>  create mode 100644 introduction.html
#>  create mode 100644 key-points.html
#>  create mode 100644 link.svg
#>  create mode 100644 md5sum.txt
#>  create mode 100644 mstile-150x150.png
#>  create mode 100644 pkgdown.css
#>  create mode 100644 pkgdown.js
#>  create mode 100644 pkgdown.yml
#>  create mode 100644 profiles.html
#>  create mode 100644 renv.lock
#>  create mode 100644 safari-pinned-tab.svg
#>  create mode 100644 site.webmanifest
#>  create mode 100644 sitemap.xml
#> Running git remote -v
#> sandpaper-local	/tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef (fetch)
#> sandpaper-local	/tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef (push)
#> Running git push --force sandpaper-local 'HEAD:gh-pages'
#> To /tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef
#>    14fa33c..cd35bcf  HEAD -> gh-pages
#> ::endgroup::
#> Running git worktree remove --force \
#>   /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/site/docs
#> Running git worktree remove --force \
#>   /tmp/Rtmpl7ps4m/file170a1776d1dc/lesson-example/site/built
tok <- Sys.time()
cli::cli_alert_info("Elapsed time: {round(tok - tik, 2)} seconds")
#> ℹ Elapsed time: 12.43 seconds

# reporting -----
# The repository and remote should both have three branches
cli::cli_h2("Local status")
#> 
#> ── Local status ──
#> 
gert::git_branch_list(repo = res)[c('name', 'commit', 'updated')]
#> # A tibble: 6 × 3
#>   name                       commit                  updated            
#>   <chr>                      <chr>                   <dttm>             
#> 1 gh-pages                   cd35bcf2e22d84d75b0598… 2023-03-09 18:20:07
#> 2 main                       608f329366172344fbb48a… 2023-03-09 18:19:54
#> 3 md-outputs                 1891d8ca88f81c4117328d… 2023-03-09 18:20:00
#> 4 sandpaper-local/gh-pages   cd35bcf2e22d84d75b0598… 2023-03-09 18:20:07
#> 5 sandpaper-local/main       608f329366172344fbb48a… 2023-03-09 18:19:54
#> 6 sandpaper-local/md-outputs 1891d8ca88f81c4117328d… 2023-03-09 18:20:00

# An indicator this worked: the first episode should be represented as
# different files across the branches:
# - main: Rmd
# - md-outputs: md
# - gh-pages: html
cli::cli_h2("First episode status")
#> ── First episode status ──
#> 
gert::git_stat_files("episodes/introduction.Rmd", repo = rmt)
#> # A tibble: 1 × 5
#>   file             created             modified            commits head 
#> * <chr>            <dttm>              <dttm>                <int> <chr>
#> 1 episodes/introd… 2023-03-09 18:19:54 2023-03-09 18:19:54       1 608f…
cli::cli_h3("rendered markdown")
#> ── rendered markdown 
gert::git_stat_files("introduction.md", repo = rmt, ref = "md-outputs")
#> # A tibble: 1 × 5
#>   file            created             modified            commits head  
#> * <chr>           <dttm>              <dttm>                <int> <chr> 
#> 1 introduction.md 2023-03-09 18:20:00 2023-03-09 18:20:00       1 1891d…
cli::cli_h3("html file")
#> 
#> ── html file 
gert::git_stat_files("introduction.html", repo = rmt, ref = "gh-pages")
#> # A tibble: 1 × 5
#>   file             created             modified            commits head 
#> * <chr>            <dttm>              <dttm>                <int> <chr>
#> 1 introduction.ht… 2023-03-09 18:20:07 2023-03-09 18:20:07       1 cd35…

# CLEAN -------------------------------------------------------------------
tik <- Sys.time()
cli::cli_h1("Clean Up")
#> 
#> ── Clean Up ────────────────────────────────────────────────────────────
snd$remove_local_remote(repo = res)
#> /tmp/Rtmpl7ps4m/REMOTE-170a36f7eaef
snd$reset_git_user(res)
# remove the test fixture and report
tryCatch(fs::dir_delete(res), error = function() FALSE)
tok <- Sys.time()
cli::cli_alert_info("Elapsed time: {round(tok - tik, 2)} seconds")
#> ℹ Elapsed time: 0.08 seconds