Internal cache for storing pre-computed lesson objects

A storage cache for pegboard::Lesson objects and other pre-computed items for use by other internal functions while sandpaper is working.

Usage

this_lesson(path)

clear_this_lesson()

set_this_lesson(path)

set_resource_list(path)

clear_resource_list(path)

Arguments

path

a path to the current lesson

Lesson Object Storage

this_lesson() will return a pegboard::Lesson object if it has previously been stored. There are three values that are cached:

.this_lesson a pegboard::Lesson object

.this_diff a charcter vector from gert::git_diff_patch()

.this_status a data frame from gert::git_status()

.this_commit the hash of the most recent commit

The function this_lesson() first checks if .this_diff is different than the output of gert::git_diff_patch(), then checks if there are any changes to gert::git_status(), and then finally checks if the commits are identical. If there are differences or the values are not previously cached, the lesson is loaded into memory, otherwise, it is fetched from the previously stored lesson.

The storage cache is in a global package object called .store, which is initialised when sandpaper is loaded via .lesson_store()

If there have been no changes git is aware of, the lesson remains the same.

Pre-Computed Object Storage

A side-effect of this_lesson() is that it will also initialise pre-computed objects that pertain to the lesson itself. These are initialised via set_globals(). These storage objects are:

.resources: a list of markdown resources for the lesson derived from get_resource_list() via set_resource_list()

this_metadata: metadata with template for including in the pages. initialised in initialise_metadata() via set_globals()

learner_globals: variables for the learner version of the pages initialised in set_globals()

instructor_globals: variables for the instructor version of the pages initialised in set_globals()

Examples

tmp <- tempfile()
create_lesson(tmp, open = FALSE)
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> ℹ No schedule set, using Rmd files in episodes/ directory.
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> → To remove this message, define your schedule in config.yaml or use `set_episodes()` to generate it.
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> ────────────────────────────────────────────────────────────────────────
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> ℹ To save this configuration, use
#> 
#> set_episodes(path = path, order = ep, write = TRUE)
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> ✔ First episode created in /tmp/RtmpBvuLuy/file1695c313e9b/episodes/introduction.Rmd
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> ℹ Workflows up-to-date!
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> ℹ Consent to use package cache provided
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> → Searching for and installing available dependencies
#> * Discovering package dependencies ... Done!
#> The following packages were discovered:
#> 
#> # ~/work/_temp/Library ===============
#> - R6            2.5.1
#> - base64enc     0.1-3
#> - bslib         0.5.0
#> - cachem        1.0.8
#> - cli           3.6.1
#> - digest        0.6.32
#> - ellipsis      0.3.2
#> - evaluate      0.21
#> - fastmap       1.1.1
#> - fontawesome   0.5.1
#> - fs            1.6.2
#> - glue          1.6.2
#> - highr         0.10
#> - htmltools     0.5.5
#> - jquerylib     0.1.4
#> - jsonlite      1.8.7
#> - knitr         1.43
#> - lifecycle     1.0.3
#> - magrittr      2.0.3
#> - memoise       2.0.1
#> - mime          0.12
#> - rappdirs      0.3.3
#> - rlang         1.1.1
#> - rmarkdown     2.23
#> - sass          0.4.6
#> - stringi       1.7.12
#> - stringr       1.5.0
#> - tinytex       0.45
#> - vctrs         0.6.3
#> - xfun          0.39
#> - yaml          2.3.7
#> 
#> They will be copied into the project library.
#> 
#> * Copying packages into the project library ... Done!
#> * Hydrated 31 packages in 0.29 seconds.
#> → Recording changes in lockfile
#> The following package(s) will be updated in the lockfile:
#> 
#> # RSPM ===============================
#> - R6            [* -> 2.5.1]
#> - base64enc     [* -> 0.1-3]
#> - bslib         [* -> 0.5.0]
#> - cachem        [* -> 1.0.8]
#> - cli           [* -> 3.6.1]
#> - digest        [* -> 0.6.32]
#> - ellipsis      [* -> 0.3.2]
#> - evaluate      [* -> 0.21]
#> - fastmap       [* -> 1.1.1]
#> - fontawesome   [* -> 0.5.1]
#> - fs            [* -> 1.6.2]
#> - glue          [* -> 1.6.2]
#> - highr         [* -> 0.10]
#> - htmltools     [* -> 0.5.5]
#> - jquerylib     [* -> 0.1.4]
#> - jsonlite      [* -> 1.8.7]
#> - knitr         [* -> 1.43]
#> - lifecycle     [* -> 1.0.3]
#> - magrittr      [* -> 2.0.3]
#> - memoise       [* -> 2.0.1]
#> - mime          [* -> 0.12]
#> - rappdirs      [* -> 0.3.3]
#> - renv          [* -> 0.17.3]
#> - rlang         [* -> 1.1.1]
#> - rmarkdown     [* -> 2.23]
#> - sass          [* -> 0.4.6]
#> - stringi       [* -> 1.7.12]
#> - stringr       [* -> 1.5.0]
#> - tinytex       [* -> 0.45]
#> - vctrs         [* -> 0.6.3]
#> - xfun          [* -> 0.39]
#> - yaml          [* -> 2.3.7]
#> 
#> The version of R recorded in the lockfile will be updated:
#> - R             [* -> 4.3.1]
#> 
#> * Lockfile written to '/tmp/RtmpBvuLuy/file1695c313e9b/renv/profiles/lesson-requirements/renv.lock'.
#> ✔ Lesson successfully created in /tmp/RtmpBvuLuy/file1695c313e9b
#> → Creating Lesson in /tmp/RtmpBvuLuy/file1695c313e9b...
#> /tmp/RtmpBvuLuy/file1695c313e9b
# Read the lesson into cache
system.time(sandpaper:::this_lesson(tmp)) 
#>    user  system elapsed 
#>   0.200   0.016   0.216 
system.time(sandpaper:::this_lesson(tmp)) # less time to read in once cached
#>    user  system elapsed 
#>   0.001   0.004   0.005 
l <- sandpaper:::this_lesson(tmp)
l
#> <Lesson>
#>   Public:
#>     blocks: function (type = NULL, level = 0, path = FALSE) 
#>     built: NULL
#>     challenges: function (path = FALSE, graph = FALSE, recurse = TRUE) 
#>     clone: function (deep = FALSE) 
#>     episodes: list
#>     extra: list
#>     files: active binding
#>     get: function (element = NULL, collection = "episodes") 
#>     handout: function (path = NULL, solution = FALSE) 
#>     initialize: function (path = ".", rmd = FALSE, jekyll = TRUE, ...) 
#>     isolate_blocks: function () 
#>     load_built: function () 
#>     n_problems: active binding
#>     path: /tmp/RtmpBvuLuy/file1695c313e9b
#>     reset: function () 
#>     rmd: FALSE
#>     sandpaper: TRUE
#>     show_problems: active binding
#>     solutions: function (path = FALSE) 
#>     summary: function (collection = "episodes") 
#>     thin: function (verbose = TRUE) 
#>     validate_divs: function () 
#>     validate_headings: function (verbose = TRUE) 
#>     validate_links: function () 
#>   Private:
#>     deep_clone: function (name, value) 
# clear the cache
sandpaper:::clear_this_lesson()
system.time(sandpaper:::this_lesson(tmp)) # have to re-read the lesson
#>    user  system elapsed 
#>   0.203   0.004   0.207 
system.time(sandpaper:::this_lesson(tmp)) 
#>    user  system elapsed 
#>   0.000   0.004   0.004 
unlink(tmp)