Set the order of items in a dropdown menu

Set the order of items in a dropdown menu

set_dropdown(path = ".", order = NULL, write = FALSE, folder)

set_config(pairs = NULL, path = ".", write = FALSE)

set_episodes(path = ".", order = NULL, write = FALSE)

set_learners(path = ".", order = NULL, write = FALSE)

set_instructors(path = ".", order = NULL, write = FALSE)

set_profiles(path = ".", order = NULL, write = FALSE)

Arguments

path
path to the lesson. Defaults to the current directory.

order
the files in the order presented (with extension)

write
if TRUE, the schedule will overwrite the schedule in the current file.

folder
one of four folders that sandpaper recognises where the files listed in order are located: episodes, learners, instructors, profiles.

pairs
a named character vector with keys as the names and the new values as the contents

Examples

tmp <- tempfile()
create_lesson(tmp, "test lesson")
#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX…
#> ✔ First episode created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX9/file12ca62b7c69b/episodes/01-introduction.Rmd
#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX…
#> ℹ Workflows up-to-date!
#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX…
#> ✔ Lesson successfully created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX9/file12ca62b7c69b
#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX…
#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX9/file12ca62b7c69b
# Change the title and License
set_config(c(title = "Absolutely Free Lesson", license = "CC0"),
  path = tmp,
  write = TRUE
)
#> ℹ Writing to /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpp1oTX9/file12ca62b7c69b/config.yaml
#> → title: 'test lesson' -> title: 'Absolutely Free Lesson'
#> → license: CC-BY 4.0 -> license: 'CC0'
create_episode("using-R", path = tmp)
print(sched <- get_episodes(tmp))
#> ℹ No schedule set, using Rmd files in episodes/ directory.
#> → To remove this message, define your schedule in config.yaml or use `set_episodes()` to generate it.
#> [1] "01-introduction.Rmd" "02-using-R.Rmd"     

# reverse the schedule
set_episodes(tmp, order = rev(sched))
#> episodes:
#> - 02-using-R.Rmd
#> - 01-introduction.Rmd
#> 
# write it
set_episodes(tmp, order = rev(sched), write = TRUE)

# see it
get_episodes(tmp)
#> [1] "02-using-R.Rmd"      "01-introduction.Rmd"