Build your lesson site

This function orchestrates rendering generated lesson content and applying the theme for the HTML site.

Usage

build_lesson(
  path = ".",
  rebuild = FALSE,
  quiet = !interactive(),
  preview = TRUE,
  override = list()
)

Arguments

path

the path to your repository (defaults to your current working directory)

rebuild

if TRUE, everything will be built from scratch as if there was no cache. Defaults to FALSE, which will only build markdown files that haven't been built before.

quiet

when TRUE, output is supressed

preview

if TRUE, the rendered website is opened in a new window

override

options to override (e.g. building to alternative paths). This is used internally and will likely be changed.

Value

TRUE if it was successful, a character vector of issues if it was unsuccessful.

Examples

tmp <- tempfile()
create_lesson(tmp, open = FALSE)
#> → Creating Lesson in /tmp/RtmpgXrbh1/file405c9b646d5...
#> ✔ First episode created in /tmp/RtmpgXrbh1/file405c9b646d5/episodes/01-introduction.Rmd
#> → Creating Lesson in /tmp/RtmpgXrbh1/file405c9b646d5...
#> ℹ Workflows up-to-date!
#> → Creating Lesson in /tmp/RtmpgXrbh1/file405c9b646d5...
#> ✔ Lesson successfully created in /tmp/RtmpgXrbh1/file405c9b646d5
#> → Creating Lesson in /tmp/RtmpgXrbh1/file405c9b646d5...
#> /tmp/RtmpgXrbh1/file405c9b646d5
create_episode("first-script", path = tmp)
check_lesson(tmp)
if (rmarkdown::pandoc_available("2.11"))
  build_lesson(tmp)