Check the lesson structure for errors

Check the lesson structure for errors

check_episode(path)

check_lesson(path = ".")

Arguments

path
the path to your lesson

Value

TRUE (invisibly) if the lesson is cromulent, otherwise, it will error with a list of things to fix.

Examples

# Everything should work out of the box
tmp <- tempfile()
create_lesson(tmp, open = FALSE)

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4…

#> ✔ First episode created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4b/file14ff125cab2f/episodes/01-introduction.Rmd

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4…

#> ℹ Workflows up-to-date!

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4…

#> ✔ Lesson successfully created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4b/file14ff125cab2f

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4…

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4b/file14ff125cab2f

check_lesson(tmp)

# if things do not work, then an error is thrown with information about 
# what has failed you
unlink(file.path(tmp, ".gitignore"))
unlink(file.path(tmp, "site"), recursive = TRUE)
try(check_lesson(tmp))

#> ! Path '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpgVDj4b/file14ff125cab2f/site' does not exist

#> ! The .gitignore file is missing the following elements:
#> episodes/*html
#> site/*
#> !site/README.md
#> .Rhistory
#> .Rapp.history
#> .RData
#> .Ruserdata
#> *-Ex.R
#> /*.tar.gz
#> /*.Rcheck/
#> .Rproj.user/
#> vignettes/*.html
#> vignettes/*.pdf
#> .httr-oauth
#> *_cache/
#> /cache/
#> *.utf8.md
#> *.knit.md
#> .Renviron
#> docs/
#> po/*~

#> ! The file 'site/README.md' does not exist

#> Error in report_validation(checklist, "There were errors with the lesson structure.") : 
#>   There were errors with the lesson structure.

unlink(tmp)