(INTERNAL) Build and deploy the site with continous integration

(INTERNAL) Build and deploy the site with continous integration

ci_deploy(
  path = ".",
  md_branch = "md-outputs",
  site_branch = "gh-pages",
  remote = "origin",
  reset = FALSE
)

Arguments

path
path to the lesson

md_branch
the branch name that contains the markdown outputs

site_branch
the branch name that contains the full HTML site

remote
the name of the git remote to which you should deploy.

reset
if TRUE, the site cache is cleared before rebuilding, this defaults to FALSE meaning that the cache in the md_branch will be used

Value

Nothing, invisibly. This is used for it's side-effect

Note

this function is not for interactive use. It requires git to be installed on your machine and will destroy anything you have in the site/ folder.