Create a carpentries lesson

This will create a boilerplate directory structure for a Carpentries lesson and initialize a git repository.

create_lesson(
  path,
  name = fs::path_file(path),
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)

Arguments

path
the path to the new lesson folder

name
the name of the lesson. If not provided, the folder name will be used.

rstudio
create an RStudio project (defaults to if RStudio exits)

open
if interactive, the lesson will open in a new editor window.

Value

the path to the new lesson

Examples

tmp <- tempfile()
on.exit(unlink(tmp))
lsn <- create_lesson(tmp, name = "This Lesson", open = FALSE)

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpPiJhc…

#> ✔ First episode created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpPiJhcl/file1a6355062c3d/episodes/01-introduction.Rmd

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpPiJhc…

#> ℹ Workflows up-to-date!

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpPiJhc…

#> ✔ Lesson successfully created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpPiJhcl/file1a6355062c3d

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpPiJhc…

lsn

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpPiJhcl/file1a6355062c3d