Test fixture functions for sandpaper

These functions are for use during testing of sandpaper and are designed to create a temporary lesson and associated remote repository (locally).

create_test_lesson()

generate_restore_fixture(tf)

setup_local_remote(
  repo,
  remote = tempfile(),
  name = "sandpaper-local",
  verbose = FALSE
)

remove_local_remote(repo, name = "sandpaper-local")

Arguments

tf
(generate_restore_fixture()) the path to the test fixture lesson

repo
path to a git repository

remote
path to an empty or uninitialized directory. Defaults to a tempfile

name
of the remote, defaults to "sandpaper-local"

verbose
if TRUE, messages and output from git will be printed to screen. Defaults to FALSE.

Value

(generate_restore_fixture) a function that will restore the test fixture

(setup_local_remote()) the repo, invisibly

(remove_local_remote()) FALSE indicating an error or a string indicating the path to the remote

Details

create_test_lesson()

This creates the test lesson and returns a function that will restore the test fixture when called with no arguments

generate_restore_fixture()

This creates the restore function for the test lesson

setup_local_remote()

Creates a local remote repository in a separate temporary folder, linked to the fixture lesson

remove_local_remote()

Destorys the local remote repository and removes it from the fixture lesson

Note

These are implemented in tests/testthat/setup.md