Setup a git worktree for concurrent manipulation of a separate branch

Setup a git worktree for concurrent manipulation of a separate branch

Usage

git_worktree_setup(
  path = ".",
  dest_dir,
  branch = "gh-pages",
  remote = "origin",
  throwaway = FALSE
)

github_worktree_commit(dir, commit_message, remote, branch)

github_worktree_remove(dir, home = NULL)

Arguments

path

path to the repository

dest_dir

path to the destination directory to contain the work tree

branch

the branch associated with the work tree (default: gh-pages)

remote

the remote name (default: origin)

throwaway

if TRUE, the worktree created is in a detached HEAD state from from the remote branch and will not create a new branch in your repository. Defaults to FALSE, which will create the branch from upstream.

Value

an expression() that calls git worktree remove on the worktree when evaluated.

Details

This function is used in continuous integration settings where we want to push derived outputs to non-main branches in our repository. We use this to populate the markdown and HTML outputs from the lesson so that we don't have to rebuild the lesson from scratch every time.

The logic behind this looks like

worktree setup
[IF BRANCH DOES NOT EXIST]
  git checkout --orphan <branch>
  git rm -rf --quiet .
  git commit --allow-empty -m
  git push remote HEAD:<branch>
  git checkout -
git fetch <remote> +refs/heads/<branch>:refs/remotes/<remote>/<branch>
git worktree add --track -B <branch> /path/to/dir <remote>/<branch>

Note

git_worktree_setup() has been modified from the logic in pkgdown::deploy_to_branch(), by Hadley Wickham.

github_worktree_commit(): Modified from pkgdown:::github_push by Hadley Wickham

github_worktree_remove(): Modified from pkgdown:::github_worktree_remove by Hadley Wickham

Examples

# Use Worktrees to deploy a lesson -----------------------------------------
# This example is a bit inovlved, but it is effectively what we do inside of
# the `ci_deploy()` function (after setting up the lesson).
#
# The setup phase will create a new lesson and a corresponding remote (self
# contained, no GitHub authentication required).
#
# The worktrees will be created for both the markdown and HTML outputs on the
# branches "md-outputs" and "gh-pages", respectively.
#
# After the worktrees are created, we will build the lesson into the
# worktrees and display the output of `git_status()` for each of the three
# branches: "main", "md-outputs", and "gh-pages"
#
# During the clean up phase, the output of `git_worktree_setup()` is
# evaluated
tik <- Sys.time()
cli::cli_h1("Set up")
#> 
#> ── Set up ──────────────────────────────────────────────────────────────
cli::cli_h2("Create Lesson")
#> 
#> ── Create Lesson ──
#> 
restore_fixture <- sandpaper:::create_test_lesson()
#> → Bootstrapping example lesson
#> → Searching for and installing available dependencies
#> * Discovering package dependencies ... Done!
#> The following packages were discovered:
#> 
#> # ~/work/_temp/Library ===============
#> - R6          2.5.1
#> - base64enc   0.1-3
#> - bslib       0.4.2
#> - cachem      1.0.7
#> - cli         3.6.0
#> - digest      0.6.31
#> - ellipsis    0.3.2
#> - evaluate    0.20
#> - fastmap     1.1.1
#> - fs          1.6.1
#> - glue        1.6.2
#> - highr       0.10
#> - htmltools   0.5.4
#> - jquerylib   0.1.4
#> - jsonlite    1.8.4
#> - knitr       1.42
#> - lifecycle   1.0.3
#> - magrittr    2.0.3
#> - memoise     2.0.1
#> - mime        0.12
#> - rappdirs    0.3.3
#> - rlang       1.0.6
#> - rmarkdown   2.20
#> - sass        0.4.5
#> - stringi     1.7.12
#> - stringr     1.5.0
#> - tinytex     0.44
#> - vctrs       0.5.2
#> - xfun        0.37
#> - yaml        2.3.7
#> 
#> They will be copied into the project library.
#> 
#> * Copying packages into the project library ... Done!
#> * Hydrated 30 packages in 0.21 seconds.
#> → Recording changes in lockfile
#> The following package(s) will be updated in the lockfile:
#> 
#> # GitHub =============================
#> - renv        [* -> rstudio/renv@HEAD]
#> 
#> # RSPM ===============================
#> - R6          [* -> 2.5.1]
#> - base64enc   [* -> 0.1-3]
#> - bslib       [* -> 0.4.2]
#> - cachem      [* -> 1.0.7]
#> - cli         [* -> 3.6.0]
#> - digest      [* -> 0.6.31]
#> - ellipsis    [* -> 0.3.2]
#> - evaluate    [* -> 0.20]
#> - fastmap     [* -> 1.1.1]
#> - fs          [* -> 1.6.1]
#> - glue        [* -> 1.6.2]
#> - highr       [* -> 0.10]
#> - htmltools   [* -> 0.5.4]
#> - jquerylib   [* -> 0.1.4]
#> - jsonlite    [* -> 1.8.4]
#> - knitr       [* -> 1.42]
#> - lifecycle   [* -> 1.0.3]
#> - magrittr    [* -> 2.0.3]
#> - memoise     [* -> 2.0.1]
#> - mime        [* -> 0.12]
#> - rappdirs    [* -> 0.3.3]
#> - rlang       [* -> 1.0.6]
#> - rmarkdown   [* -> 2.20]
#> - sass        [* -> 0.4.5]
#> - stringi     [* -> 1.7.12]
#> - stringr     [* -> 1.5.0]
#> - tinytex     [* -> 0.44]
#> - vctrs       [* -> 0.5.2]
#> - xfun        [* -> 0.37]
#> - yaml        [* -> 2.3.7]
#> 
#> The version of R recorded in the lockfile will be updated:
#> - R           [* -> 4.2.2]
#> 
#> * Lockfile written to '/tmp/RtmphQGbTg/file17067e477a54/lesson-example/renv/profiles/lesson-requirements/renv.lock'.
res <- getOption("sandpaper.test_fixture")
sandpaper:::check_git_user(res)
cli::cli_h2("Create Remote")
#> 
#> ── Create Remote ──
#> 
rmt <- fs::file_temp(pattern = "REMOTE-")
sandpaper:::setup_local_remote(repo = res, remote = rmt, verbose = FALSE)
tok <- Sys.time()
cli::cli_alert_info("Elapsed time: {round(tok - tik, 2)} seconds")
#> ℹ Elapsed time: 3.05 seconds
tik <- Sys.time()
cli::cli_h2("Create Worktrees")
#> 
#> ── Create Worktrees ──
#> 
db <- sandpaper:::git_worktree_setup(res, fs::path(res, "site", "built"),
  branch = "md-outputs", remote = "sandpaper-local"
)
#> ::group::Create New Branch
#> Running git checkout --orphan md-outputs
#> Switched to a new branch 'md-outputs'
#> Running git rm -rf --quiet .
#> Running git commit --allow-empty -m 'Initializing md-outputs branch'
#> [md-outputs (root-commit) 389e969] Initializing md-outputs branch
#> Running git push sandpaper-local 'HEAD:md-outputs'
#> To /tmp/RtmphQGbTg/REMOTE-17066169b822
#>  * [new branch]      HEAD -> md-outputs
#> Running git checkout main
#> Your branch is up to date with 'sandpaper-local/main'.
#> Switched to branch 'main'
#> ::endgroup::
#> ::group::Fetch sandpaper-local/md-outputs
#> Running git remote set-branches sandpaper-local md-outputs
#> Running git fetch sandpaper-local md-outputs
#> From /tmp/RtmphQGbTg/REMOTE-17066169b822
#>  * branch            md-outputs -> FETCH_HEAD
#> Running git remote set-branches sandpaper-local '*'
#> ::endgroup::
#> ::group::Add worktree for sandpaper-local/md-outputs in site/built
#> Running git worktree add --track -B md-outputs \
#>   /tmp/RtmphQGbTg/file17067e477a54/lesson-example/site/built \
#>   sandpaper-local/md-outputs
#> Preparing worktree (resetting branch 'md-outputs'; was at 389e969)
#> branch 'md-outputs' set up to track 'sandpaper-local/md-outputs'.
#> HEAD is now at 389e969 Initializing md-outputs branch
#> ::endgroup::
ds <- sandpaper:::git_worktree_setup(res, fs::path(res, "site", "docs"),
  branch = "gh-pages", remote = "sandpaper-local"
)
#> ::group::Create New Branch
#> Running git checkout --orphan gh-pages
#> Switched to a new branch 'gh-pages'
#> Running git rm -rf --quiet .
#> Running git commit --allow-empty -m 'Initializing gh-pages branch'
#> [gh-pages (root-commit) e909926] Initializing gh-pages branch
#> Running git push sandpaper-local 'HEAD:gh-pages'
#> To /tmp/RtmphQGbTg/REMOTE-17066169b822
#>  * [new branch]      HEAD -> gh-pages
#> Running git checkout main
#> Your branch is up to date with 'sandpaper-local/main'.
#> Switched to branch 'main'
#> ::endgroup::
#> ::group::Fetch sandpaper-local/gh-pages
#> Running git remote set-branches sandpaper-local gh-pages
#> Running git fetch sandpaper-local gh-pages
#> From /tmp/RtmphQGbTg/REMOTE-17066169b822
#>  * branch            gh-pages   -> FETCH_HEAD
#> Running git remote set-branches sandpaper-local '*'
#> ::endgroup::
#> ::group::Add worktree for sandpaper-local/gh-pages in site/docs
#> Running git worktree add --track -B gh-pages \
#>   /tmp/RtmphQGbTg/file17067e477a54/lesson-example/site/docs \
#>   sandpaper-local/gh-pages
#> Preparing worktree (resetting branch 'gh-pages'; was at e909926)
#> branch 'gh-pages' set up to track 'sandpaper-local/gh-pages'.
#> HEAD is now at e909926 Initializing gh-pages branch
#> ::endgroup::
tok <- Sys.time()
cli::cli_alert_info("Elapsed time: {round(tok - tik, 2)} seconds")
#> ℹ Elapsed time: 0.31 seconds
tik <- Sys.time()
cli::cli_h1("Build Lesson into worktrees")
#> 
#> ── Build Lesson into worktrees ─────────────────────────────────────────
build_lesson(res, quiet = TRUE, preview = FALSE)
cli::cli_h2("git status: {gert::git_branch(repo = res)}")
#> 
#> ── git status: main ──
#> 
print(gert::git_status(repo = res))
#> # A tibble: 0 × 3
#> # … with 3 variables: file <chr>, status <chr>, staged <lgl>
cli::cli_h2('git status: {gert::git_branch(repo = fs::path(res, "site", "built"))}')
#> ── git status: md-outputs ──
#> 
print(gert::git_status(repo = fs::path(res, "site", "built")))
#> # A tibble: 12 × 3
#>    file                status staged
#>    <chr>               <chr>  <lgl> 
#>  1 CODE_OF_CONDUCT.md  new    FALSE 
#>  2 LICENSE.md          new    FALSE 
#>  3 config.yaml         new    FALSE 
#>  4 fig/                new    FALSE 
#>  5 index.md            new    FALSE 
#>  6 instructor-notes.md new    FALSE 
#>  7 introduction.md     new    FALSE 
#>  8 learner-profiles.md new    FALSE 
#>  9 links.md            new    FALSE 
#> 10 md5sum.txt          new    FALSE 
#> 11 renv.lock           new    FALSE 
#> 12 setup.md            new    FALSE 
cli::cli_h2('git status: {gert::git_branch(repo = fs::path(res, "site", "docs"))}')
#> ── git status: gh-pages ──
#> 
print(gert::git_status(repo = fs::path(res, "site", "docs")))
#> # A tibble: 34 × 3
#>    file                       status staged
#>    <chr>                      <chr>  <lgl> 
#>  1 .nojekyll                  new    FALSE 
#>  2 CODE_OF_CONDUCT.html       new    FALSE 
#>  3 LICENSE.html               new    FALSE 
#>  4 aio.html                   new    FALSE 
#>  5 android-chrome-192x192.png new    FALSE 
#>  6 android-chrome-512x512.png new    FALSE 
#>  7 apple-touch-icon.png       new    FALSE 
#>  8 assets/                    new    FALSE 
#>  9 bootstrap-toc.css          new    FALSE 
#> 10 bootstrap-toc.js           new    FALSE 
#> # … with 24 more rows
tok <- Sys.time()
cli::cli_alert_info("Elapsed time: {round(tok - tik, 2)} seconds")
#> ℹ Elapsed time: 9.36 seconds
tik <- Sys.time()
cli::cli_h1("Clean Up")
#> 
#> ── Clean Up ────────────────────────────────────────────────────────────
cli::cli_alert_info("object db is an expression that evaluates to {.code {db}}")
#> ℹ object db is an expression that evaluates to `sandpaper:::github_worktree_remove("/tmp/RtmphQGbTg/file17067e477a54/lesson-example/site/built", "/tmp/RtmphQGbTg/file17067e477a54/lesson-example")`
eval(db)
#> Running git worktree remove --force \
#>   /tmp/RtmphQGbTg/file17067e477a54/lesson-example/site/built
#> $status
#> [1] 0
#> 
#> $stdout
#> [1] ""
#> 
#> $stderr
#> [1] ""
#> 
#> $timeout
#> [1] FALSE
#> 
cli::cli_alert_info("object ds is an expression that evaluates to {.code {ds}}")
#> ℹ object ds is an expression that evaluates to `sandpaper:::github_worktree_remove("/tmp/RtmphQGbTg/file17067e477a54/lesson-example/site/docs", "/tmp/RtmphQGbTg/file17067e477a54/lesson-example")`
eval(ds)
#> Running git worktree remove --force \
#>   /tmp/RtmphQGbTg/file17067e477a54/lesson-example/site/docs
#> $status
#> [1] 0
#> 
#> $stdout
#> [1] ""
#> 
#> $stderr
#> [1] ""
#> 
#> $timeout
#> [1] FALSE
#> 
sandpaper:::remove_local_remote(repo = res)
#> /tmp/RtmphQGbTg/REMOTE-17066169b822
sandpaper:::reset_git_user(res)
# remove the test fixture and report
tryCatch(fs::dir_delete(res), error = function() FALSE)
tok <- Sys.time()
cli::cli_alert_info("Elapsed time: {round(tok - tik, 2)} seconds")
#> ℹ Elapsed time: 0.12 seconds