Internal cache for storing lesson objects

A storage cache for pegboard::Lesson objects that works across the functions while sandpaper is working.

set_this_lesson(path)

clear_this_lesson()

this_lesson(path)

Arguments

path
a path to the current lesson

Details

this_lesson() will return a pegboard::Lesson object if it has previously been stored. There are three values that are cached:

.this_lesson a pegboard::Lesson object

.this_diff a charcter vector from gert::git_diff_patch()

.this_status a data frame from gert::git_status()

The function this_lesson() first checks if .this_diff is different than the output of gert::git_diff_patch() and then check if there are any changes to gert::git_status(). If there are no differences or the values are not previously cached, the lesson is loaded into memory.

If there have been no changes git is aware of, the lesson remains the same.

Examples

tmp <- tempfile()
create_lesson(tmp, open = FALSE)

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp9ayZ4…

#> ✔ First episode created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp9ayZ4n/file1ec71e7f7a6e/episodes/01-introduction.Rmd

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp9ayZ4…

#> ℹ Workflows up-to-date!

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp9ayZ4…

#> ✔ Lesson successfully created in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp9ayZ4n/file1ec71e7f7a6e

#> → Creating Lesson in /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp9ayZ4…

#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp9ayZ4n/file1ec71e7f7a6e

# Read the lesson into cache
system.time(sandpaper:::this_lesson(tmp)) 

#>    user  system elapsed 
#>   0.153   0.022   0.255 

system.time(sandpaper:::this_lesson(tmp)) # less time to read in once cached

#>    user  system elapsed 
#>   0.002   0.005   0.008 

l <- sandpaper:::this_lesson(tmp)
l

#> <Lesson>
#>   Public:
#>     blocks: function (type = NULL, level = 0, path = FALSE) 
#>     challenges: function (path = FALSE, graph = FALSE, recurse = TRUE) 
#>     clone: function (deep = FALSE) 
#>     episodes: list
#>     extra: list
#>     files: active binding
#>     handout: function (path = NULL, solution = FALSE) 
#>     initialize: function (path = ".", rmd = FALSE, jekyll = TRUE, ...) 
#>     isolate_blocks: function () 
#>     n_problems: active binding
#>     path: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmp9a ...
#>     reset: function () 
#>     rmd: FALSE
#>     show_problems: active binding
#>     solutions: function (path = FALSE) 
#>     thin: function (verbose = TRUE) 
#>     validate_headings: function (verbose = TRUE) 
#>     validate_links: function () 
#>   Private:
#>     deep_clone: function (name, value) 

# clear the cache
sandpaper:::clear_this_lesson()
system.time(sandpaper:::this_lesson(tmp)) # have to re-read the lesson

#>    user  system elapsed 
#>   0.135   0.015   0.151 

system.time(sandpaper:::this_lesson(tmp)) 

#>    user  system elapsed 
#>   0.003   0.007   0.010 

unlink(tmp)