Establish and Manage Translation Strings

Establish and Manage Translation Strings

Usage

these

establish_translation_vars()

set_language(lang = NULL)

Format

An object of class environment of length 1.

Arguments

lang

a two-letter language code (optionally with a country code). Defaults to NULL, which falls back to English (the language of the source code).

Details

A page generated from sandpaper is made up of user-provided content wrapped into templated HTML provided by varnish. Since users can provide content in any human language that can be processed by computer, sandpaper must have a method to match the template content to the language of the content. These translations are added to the source of sandpaper by volunteers using the potools package.

Template content strings are stored in the $translations element of the these global environment variable, which is generated by the establish_translation_vars() function, which is run every time sandpaper is loaded. The $translations element consists of three lists:

varnish: a list of translated strings that are passed directly to varnish

computed: a list of translated strings that are incorporated into the HTML content before it is passed to varnish.

src: a source list of the above two lists that serves as the source of the translations.

Whenever a lesson is built, the function set_language() will read the lang item from the config.yaml and apply translations to these$translations$varnish and these$translations$computed before the lesson is generated or updated.

List of Translation Variables

There are 62 translations generated by set_language() that correspond to the following variables in varnish:

variable	string

translate.SkipToMain	'Skip to main content'

translate.iPreAlpha	'Pre-Alpha'

translate.PreAlphaNote	'This lesson is in the pre-alpha phase, which means that it is in early development, but has not yet been taught.'

translate.AlphaNote	'This lesson is in the alpha phase, which means that it has been taught once and lesson authors are iterating on feedback.'

translate.iAlpha	'Alpha'

translate.BetaNote	'This lesson is in the beta phase, which means that it is ready for teaching by instructors outside of the original author team.'

translate.iBeta	'Beta'

translate.PeerReview	'This lesson has passed peer review.'

translate.InstructorView	'Instructor View'

translate.LearnerView	'Learner View'

translate.MainNavigation	'Main Navigation'

translate.ToggleNavigation	'Toggle Navigation'

translate.Menu	'Menu'

translate.SearchButton	'search button'

translate.Setup	'Setup'

translate.KeyPoints	'Key Points'

translate.InstructorNotes	'Instructor Notes'

translate.Glossary	'Glossary'

translate.LearnerProfiles	'Learner Profiles'

translate.More	'More'

translate.Search	'Search'

translate.LessonProgress	'Lesson Progress'

translate.CloseMenu	'close menu'

translate.EPISODES	'EPISODES'

translate.Home	'Home'

translate.HomePageNav	'Home Page Navigation'

translate.RESOURCES	'RESOURCES'

translate.ExtractAllImages	'Extract All Images'

translate.AIO	'See all in one page'

translate.DownloadHandout	'Download Lesson Handout'

translate.ExportSlides	'Export Chapter Slides'

translate.PreviousAndNext	'Previous and Next Chapter'

translate.Previous	'Previous'

translate.EstimatedTime	'Estimated time: {icons$clock} {minutes} minutes'

translate.Next	'Next'

translate.NextChapter	'Next Chapter'

translate.LastUpdate	'Last updated on {updated}'

translate.EditThisPage	'Edit this page'

translate.ExpandAllSolutions	'Expand All Solutions'

translate.SetupInstructions	'Setup Instructions'

translate.DownloadFiles	'Download files required for the lesson'

translate.ActualScheduleNote	'The actual schedule may vary slightly depending on the topics and exercises chosen by the instructor.'

translate.BackToTop	'Back To Top'

translate.SpanToTop	'<(Back)> To Top'

translate.ThisLessonCoC	'This lesson is subject to the <(Code of Conduct)>'

translate.CoC	'Code of Conduct'

translate.EditOnGH	'Edit on GitHub'

translate.Contributing	'Contributing'

translate.Source	'Source'

translate.Cite	'Cite'

translate.Contact	'Contact'

translate.About	'About'

translate.MaterialsLicensedUnder	'Materials licensed under {license} by {authors}'

translate.TemplateLicense	'Template licensed under <(CC-BY 4.0)> by {template_authors}'

translate.Carpentries	'The Carpentries'

translate.BuiltWith	'Built with {sandpaper_link}, {pegboard_link}, and {varnish_link}'

translate.ExpandAllSolutions	'Expand All Solutions'

translate.CollapseAllSolutions	'Collapse All Solutions'

translate.Collapse	'Collapse'

translate.Episodes	'Episodes'

translate.GiveFeedback	'Give Feedback'

translate.LearnMore	'Learn More'

In addition, there are 26 translations that are inserted before they get to varnish:

variable	string

OUTPUT	'OUTPUT'

WARNING	'WARNING'

ERROR	'ERROR'

Overview	'Overview'

Questions	'Questions'

Objectives	'Objectives'

Callout	'Callout'

Challenge	'Challenge'

Prereq	'Prerequisite'

Checklist	'Checklist'

Discussion	'Discussion'

Testimonial	'Testimonial'

Keypoints	'Key Points'

Show me the solution	'Show me the solution'

Give me a hint	'Give me a hint'

Show details	'Show details'

Instructor Note	'Instructor Note'

SummaryAndSetup	'Summary and Setup'

SummaryAndSchedule	'Summary and Schedule'

AllInOneView	'All in One View'

PageNotFound	'Page not found'

AllImages	'All Images'

Anchor	'anchor'

Figure	'Figure {element}'

ImageOf	'Image {i} of {n}: {sQuote(txt)}'

Finish	'Finish'

Examples

snd <- asNamespace("sandpaper")
snd$these$translations$varnish
#> $SkipToMain
#> [1] "Skip to main content"
#> 
#> $iPreAlpha
#> [1] "Pre-Alpha"
#> 
#> $PreAlphaNote
#> [1] "This lesson is in the pre-alpha phase, which means that it is in early development, but has not yet been taught."
#> 
#> $AlphaNote
#> [1] "This lesson is in the alpha phase, which means that it has been taught once and lesson authors are iterating on feedback."
#> 
#> $iAlpha
#> [1] "Alpha"
#> 
#> $BetaNote
#> [1] "This lesson is in the beta phase, which means that it is ready for teaching by instructors outside of the original author team."
#> 
#> $iBeta
#> [1] "Beta"
#> 
#> $PeerReview
#> [1] "This lesson has passed peer review."
#> 
#> $InstructorView
#> [1] "Instructor View"
#> 
#> $LearnerView
#> [1] "Learner View"
#> 
#> $MainNavigation
#> [1] "Main Navigation"
#> 
#> $ToggleNavigation
#> [1] "Toggle Navigation"
#> 
#> $Menu
#> [1] "Menu"
#> 
#> $SearchButton
#> [1] "search button"
#> 
#> $Setup
#> [1] "Setup"
#> 
#> $KeyPoints
#> [1] "Key Points"
#> 
#> $InstructorNotes
#> [1] "Instructor Notes"
#> 
#> $Glossary
#> [1] "Glossary"
#> 
#> $LearnerProfiles
#> [1] "Learner Profiles"
#> 
#> $More
#> [1] "More"
#> 
#> $Search
#> [1] "Search"
#> 
#> $LessonProgress
#> [1] "Lesson Progress"
#> 
#> $CloseMenu
#> [1] "close menu"
#> 
#> $EPISODES
#> [1] "EPISODES"
#> 
#> $Home
#> [1] "Home"
#> 
#> $HomePageNav
#> [1] "Home Page Navigation"
#> 
#> $RESOURCES
#> [1] "RESOURCES"
#> 
#> $ExtractAllImages
#> [1] "Extract All Images"
#> 
#> $AIO
#> [1] "See all in one page"
#> 
#> $DownloadHandout
#> [1] "Download Lesson Handout"
#> 
#> $ExportSlides
#> [1] "Export Chapter Slides"
#> 
#> $PreviousAndNext
#> [1] "Previous and Next Chapter"
#> 
#> $Previous
#> [1] "Previous"
#> 
#> $EstimatedTime
#> [1] "Estimated time: {icons$clock} {minutes} minutes"
#> 
#> $Next
#> [1] "Next"
#> 
#> $NextChapter
#> [1] "Next Chapter"
#> 
#> $LastUpdate
#> [1] "Last updated on {updated}"
#> 
#> $EditThisPage
#> [1] "Edit this page"
#> 
#> $ExpandAllSolutions
#> [1] "Expand All Solutions"
#> 
#> $SetupInstructions
#> [1] "Setup Instructions"
#> 
#> $DownloadFiles
#> [1] "Download files required for the lesson"
#> 
#> $ActualScheduleNote
#> [1] "The actual schedule may vary slightly depending on the topics and exercises chosen by the instructor."
#> 
#> $BackToTop
#> [1] "Back To Top"
#> 
#> $SpanToTop
#> [1] "<(Back)> To Top"
#> 
#> $ThisLessonCoC
#> [1] "This lesson is subject to the <(Code of Conduct)>"
#> 
#> $CoC
#> [1] "Code of Conduct"
#> 
#> $EditOnGH
#> [1] "Edit on GitHub"
#> 
#> $Contributing
#> [1] "Contributing"
#> 
#> $Source
#> [1] "Source"
#> 
#> $Cite
#> [1] "Cite"
#> 
#> $Contact
#> [1] "Contact"
#> 
#> $About
#> [1] "About"
#> 
#> $MaterialsLicensedUnder
#> [1] "Materials licensed under {license} by {authors}"
#> 
#> $TemplateLicense
#> [1] "Template licensed under <(CC-BY 4.0)> by {template_authors}"
#> 
#> $Carpentries
#> [1] "The Carpentries"
#> 
#> $BuiltWith
#> [1] "Built with {sandpaper_link}, {pegboard_link}, and {varnish_link}"
#> 
#> $ExpandAllSolutions
#> [1] "Expand All Solutions"
#> 
#> $CollapseAllSolutions
#> [1] "Collapse All Solutions"
#> 
#> $Collapse
#> [1] "Collapse"
#> 
#> $Episodes
#> [1] "Episodes"
#> 
#> $GiveFeedback
#> [1] "Give Feedback"
#> 
#> $LearnMore
#> [1] "Learn More"
#> 
snd$these$translations$computed
#> $OUTPUT
#> [1] "OUTPUT"
#> 
#> $WARNING
#> [1] "WARNING"
#> 
#> $ERROR
#> [1] "ERROR"
#> 
#> $Overview
#> [1] "Overview"
#> 
#> $Questions
#> [1] "Questions"
#> 
#> $Objectives
#> [1] "Objectives"
#> 
#> $Callout
#> [1] "Callout"
#> 
#> $Challenge
#> [1] "Challenge"
#> 
#> $Prereq
#> [1] "Prerequisite"
#> 
#> $Checklist
#> [1] "Checklist"
#> 
#> $Discussion
#> [1] "Discussion"
#> 
#> $Testimonial
#> [1] "Testimonial"
#> 
#> $Keypoints
#> [1] "Key Points"
#> 
#> $`Show me the solution`
#> [1] "Show me the solution"
#> 
#> $`Give me a hint`
#> [1] "Give me a hint"
#> 
#> $`Show details`
#> [1] "Show details"
#> 
#> $`Instructor Note`
#> [1] "Instructor Note"
#> 
#> $SummaryAndSetup
#> [1] "Summary and Setup"
#> 
#> $SummaryAndSchedule
#> [1] "Summary and Schedule"
#> 
#> $AllInOneView
#> [1] "All in One View"
#> 
#> $PageNotFound
#> [1] "Page not found"
#> 
#> $AllImages
#> [1] "All Images"
#> 
#> $Anchor
#> [1] "anchor"
#> 
#> $Figure
#> [1] "Figure {element}"
#> 
#> $ImageOf
#> [1] "Image {i} of {n}: {sQuote(txt)}"
#> 
#> $Finish
#> [1] "Finish"
#> 
snd$set_language("es")
snd$these$translations$computed
#> $OUTPUT
#> [1] "SALIDA"
#> 
#> $WARNING
#> [1] "ADVERTENCIA"
#> 
#> $ERROR
#> [1] "ERROR"
#> 
#> $Overview
#> [1] "Hoja de ruta"
#> 
#> $Questions
#> [1] "Preguntas"
#> 
#> $Objectives
#> [1] "Objetivos"
#> 
#> $Callout
#> [1] "Aviso"
#> 
#> $Challenge
#> [1] "Desafío"
#> 
#> $Prereq
#> [1] "Pre requisito"
#> 
#> $Checklist
#> [1] "Lista de verificación"
#> 
#> $Discussion
#> [1] "Discusión"
#> 
#> $Testimonial
#> [1] "Testimonio"
#> 
#> $Keypoints
#> [1] "Puntos Clave"
#> 
#> $`Show me the solution`
#> [1] "Mostrar la solución"
#> 
#> $`Give me a hint`
#> [1] "Dar una pista"
#> 
#> $`Show details`
#> [1] "Mostrar detalles"
#> 
#> $`Instructor Note`
#> [1] "Notas de la instructora"
#> 
#> $SummaryAndSetup
#> [1] "Resumen y configuración"
#> 
#> $SummaryAndSchedule
#> [1] "Resumen y agenda"
#> 
#> $AllInOneView
#> [1] "Todo en una sola página"
#> 
#> $PageNotFound
#> [1] "Página no encontrada"
#> 
#> $AllImages
#> [1] "Todas las imágenes"
#> 
#> $Anchor
#> [1] "anclaje"
#> 
#> $Figure
#> [1] "Figura {element}"
#> 
#> $ImageOf
#> [1] "Imagen {i} de {n}: {sQuote(txt)}"
#> 
#> $Finish
#> [1] "Final"
#> 
snd$set_language() # defaults to English
snd$these$translations$computed
#> $OUTPUT
#> [1] "OUTPUT"
#> 
#> $WARNING
#> [1] "WARNING"
#> 
#> $ERROR
#> [1] "ERROR"
#> 
#> $Overview
#> [1] "Overview"
#> 
#> $Questions
#> [1] "Questions"
#> 
#> $Objectives
#> [1] "Objectives"
#> 
#> $Callout
#> [1] "Callout"
#> 
#> $Challenge
#> [1] "Challenge"
#> 
#> $Prereq
#> [1] "Prerequisite"
#> 
#> $Checklist
#> [1] "Checklist"
#> 
#> $Discussion
#> [1] "Discussion"
#> 
#> $Testimonial
#> [1] "Testimonial"
#> 
#> $Keypoints
#> [1] "Key Points"
#> 
#> $`Show me the solution`
#> [1] "Show me the solution"
#> 
#> $`Give me a hint`
#> [1] "Give me a hint"
#> 
#> $`Show details`
#> [1] "Show details"
#> 
#> $`Instructor Note`
#> [1] "Instructor Note"
#> 
#> $SummaryAndSetup
#> [1] "Summary and Setup"
#> 
#> $SummaryAndSchedule
#> [1] "Summary and Schedule"
#> 
#> $AllInOneView
#> [1] "All in One View"
#> 
#> $PageNotFound
#> [1] "Page not found"
#> 
#> $AllImages
#> [1] "All Images"
#> 
#> $Anchor
#> [1] "anchor"
#> 
#> $Figure
#> [1] "Figure {element}"
#> 
#> $ImageOf
#> [1] "Image {i} of {n}: {sQuote(txt)}"
#> 
#> $Finish
#> [1] "Finish"
#>